Automate Dependency Install and Terminal Configuration
The dependency_install.sh script determines the system distro and then
downloads the correct packages. As of now the supported distros include
Fedora, RHEL, CentOS, Ubuntu, and Debian. The script then checks the
terminal language requirements. If either one of these requirements are
not met the program will download a compatible version respectively. The
script will then prepare an ssh key for the user to link to their IBM
GitHub Enterprise account before finally listing any packages that could
not be installed.

Signed-off-by: Ram Pai <linuxram@us.ibm.com>
[Reformatted the description]
[fix total cpu calculation logic]
[fix dependency on i686 packages]
Signed-off-by: Rohan Pankaj <Rohan.Pankaj@ibm.com>
Signed-off-by: Sethu Eapen <Sethu.Eapen@ibm.com>